Move calendar functions to a Calendar API
*Created a Calendar.API project

Required Reviewer(es)

Optional Reviewer(es)

This change addresses the need by:

*
