+++ Calendar.API/paket.references
@@ -1,0 +1,4 @@
+Google.Apis.Calendar.v3
+Google.Apis.Discovery.v1
+Microsoft.Extensions.Configuration
+Microsoft.Extensions.Configuration.Json
