+++ ServiceScheduler/ServiceScheduler.csproj
@@ -10,5 +10,8 @@
     </None>
     <None Update="credentials.json" CopyToOutputDirectory="PreserveNewest" />
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Calendar.API\Calendar.API.csproj" />
+  </ItemGroup>
   <Import Project="..\.paket\Paket.Restore.targets" />
 </Project>
